requirements/test.txt
-r base.txt

pytest
pytest-mock
pytest-django
pytest-cov
pytest-xdist
pytest-selenium
docker-py
ddt
python-coveralls==2.9.1
mock==1.0.1
behave-django==0.3.0
factory-boy==2.8.1
django-nose==1.4.4
codecov==2.0.9
coverage==4.0.3
pycodestyle
diff-cover
pytest-metadata==1.5.1
